+++ src/irmin-pack/inode_intf.ml
@@ -64,6 +64,9 @@ module type Value = sig
        and type metadata := metadata
 
   val nb_children : t -> int
+
+  val recompute_hash : t -> hash
+  (** Recompute hash for inodes, used in eager integrity checks.*)
 end
 
 module type Raw = sig