+++ CHANGES.md
@@ -1,3 +1,9 @@
+## Unreleased
+
+### Changed
+  - Improve GC reachability traversal to optimize memory, speed and remove
+    the need for temporary files. (#2085, @art-w)
+
 ## 3.5.0
 
 ### Added