+++ CHANGES.md
@@ -1,3 +1,11 @@
+## Unreleased
+
+### Fixed
+
+- **irmin-pack**
+  - Integrity check of a commit works on stores using the minimal indexing
+    strategy. (#2160, @icristescu)
+
 ## 3.5.0
 
 ### Added
@@ -22,8 +30,6 @@
   - Modify GC to work with new chunked suffix. See `examples/gc.ml` for a
     demonstration of how it works with the new `split` function. (#2126,
     @metanivek)
-
-### Fixed
 
 ## 3.4.3 (2022-10-19)